Move all of FoundDefinitions to core
