+++ core/FoundDefinitions.cc
@@ -1,0 +1,5 @@
+#include "core/FoundDefinitions.h"
+
+using namespace std;
+
+namespace sorbet::core {} // namespace sorbet::core
